干焦粗坯用會著的就囥佇粗坯
