+++ 臺灣言語工具/基本物件/公用變數.py
@@ -1,11 +1,10 @@
 # -*- coding: utf-8 -*-
 # 瀏覽器希望無音愛有空白，但是處理標音時希望是佮好認的
 from 臺灣言語工具.音標系統.客話.臺灣客家話拼音 import 臺灣客家話拼音調類對照表
-import re
+from 臺灣言語工具.音標系統.閩南語.臺灣閩南語羅馬字拼音轉方音符號吳守禮改良式模組 import 臺灣閩南語羅馬字拼音對照吳守禮方音聲調表
 import unicodedata
 
 
-from 臺灣言語工具.音標系統.閩南語.臺灣閩南語羅馬字拼音轉方音符號吳守禮改良式模組 import 臺灣閩南語羅馬字拼音對照吳守禮方音聲調表
 無音 = ''  # '　'
 # sui1 koo1-niu5 =>　媠　姑娘
 分字符號 = '-'
@@ -46,7 +45,6 @@ from 臺灣言語工具.音標系統.閩南語.臺灣閩南語羅馬字拼音轉
 
 # Ll　小寫， Lu　大寫， Md　數字， Mn　有調號英文，Lo　其他, So 組字式符號…
 _統一碼羅馬字類 = {'Ll', 'Lu', 'Mn'}
-統一碼漢字佮組字式類 = {'Lo', 'So'}
 
 
 def 敢是拼音字元(字元, 種類=None):
@@ -62,15 +60,3 @@ def 敢是拼音字元(字元, 種類=None):
 
 def 敢是注音符號(字元):
     return unicodedata.name(字元, '').startswith('BOPOMOFO LETTER')
-
-
-_hiragana範圍 = re.compile(r'[ぁ-ゟ]\Z')
-_katakana範圍 = re.compile(r'[゠-ヿ]\Z')
-
-
-def 敢是hiragana(字元):
-    return _hiragana範圍.match(字元)
-
-
-def 敢是katakana(字元):
-    return _katakana範圍.match(字元)
